Created AlexDemoScene.tscn and MoneyManager.gd + added as autoload
My demo scene is just a copy of Eyan's environment node (bless up my friend). I'm making this push in the event that something comes up and I'm unable to continue, but as far as I've checked this should be the setup for the skeleton. The next step is to actually make the money object and have a click event listener or whatever the godot equivalent is.