Added primitive coin, setup for camera click events.
Made a material for a coin, easily adjustable later. Something to note is that the capsule collision object (which i read up on stackexchange) is scaled to fit the coin, but gives a warning that it not being uniform might not perform as expected. I'm keeping it for now, but if it comes to it i'll look into it.