Adjusted Starting Money and Tower Damage
Made some tweaks and play tested to what I believe is a reasonable start and spending economy for the enemies presented.
Made to be adjustable as we move forward with more concrete enemy types and waves and differently evolving attributes. Upgrades will take more money into account.
In the event that someone wants to change the money themselves, they are free to do so within the MoneyManager.gd file, as it is singleton and autoloaded into any scene for ease. Can turn into node upon request.